Replace recognition feedback drawer with inline answer reveal.
Swap the input controls for an in-flow feedback panel instead of opening a bottom sheet, so touch sessions avoid the drawer overlay and keyboard layout churn.
